code/requirements.txt
sentence_transformers==2.2.2
elasticsearch==8.8.0
openai==1.7.2
python-dotenv==1.0.0
